APP-2539 healthCreditsRedeemed should be float

+++ packages/api/src/profiles-service/resolvers/getMedicineOrdersOMSList.ts
@@ -121,7 +121,7 @@ export const getMedicineOrdersOMSListTypeDefs = gql`
     responseCode: String
     responseMessage: String
     bankTxnId: String
-    healthCreditsRedeemed: Int
+    healthCreditsRedeemed: Float
     healthCreditsRedemptionRequest: BlockUserPointsResponse
   }
